Clean up old aframe components
