--- src/components/tools/pen-laser.js
@@ -1,108 +1,0 @@
-const InterpolationBuffer = require("buffered-interpolation");
-import { convertStandardMaterial } from "../../utils/material-utils";
-
-function almostEquals(epsilon, u, v) {
-    return Math.abs(u.x - v.x) < epsilon && Math.abs(u.y - v.y) < epsilon && Math.abs(u.z - v.z) < epsilon;
-}
-
-AFRAME.registerComponent("pen-laser", {
-    schema: {
-        color: { type: "color", default: "#FF0033" },
-        laserVisible: { default: false },
-        laserInHand: { default: false },
-        laserOrigin: { default: { x: 0, y: 0, z: 0 }, type: "vec3" },
-        remoteLaserOrigin: { default: { x: 0, y: 0, z: 0 }, type: "vec3" },
-        laserTarget: { default: { x: 0, y: 0, z: 0 }, type: "vec3" }
-    },
-
-    init() {
-        let material = new THREE.MeshStandardMaterial({ color: "red", opacity: 0.5, transparent: true, visible: true });
-        const quality = window.APP.store.state.preferences.materialQualitySetting;
-        material = convertStandardMaterial(material, quality);
-
-        const lineCurve = new THREE.LineCurve3(new THREE.Vector3(0, 0, 0), new THREE.Vector3(0, 0, 2));
-        const geometry = new THREE.TubeBufferGeometry(lineCurve, 2, 0.003, 8, true);
-        this.laser = new THREE.Mesh(geometry, material);
-
-        //prevents the line from being a raycast target for the cursor
-        this.laser.raycast = function () {};
-
-        this.el.sceneEl.setObject3D(`pen-laser-${this.laser.uuid}`, this.laser);
-
-        this.originBuffer = new InterpolationBuffer(InterpolationBuffer.MODE_LERP, 0.1);
-        this.targetBuffer = new InterpolationBuffer(InterpolationBuffer.MODE_LERP, 0.1);
-    },
-
-    update: (() => {
-        const originBufferPosition = new THREE.Vector3();
-        const targetBufferPosition = new THREE.Vector3();
-
-        return function (prevData) {
-            if (prevData.color != this.data.color) {
-                this.laser.material.color.set(this.data.color);
-            }
-
-            if (
-                prevData.remoteLaserOrigin &&
-                !almostEquals(0.001, prevData.remoteLaserOrigin, this.data.remoteLaserOrigin)
-            ) {
-                this.originBuffer.setPosition(
-                    originBufferPosition.set(
-                        this.data.remoteLaserOrigin.x,
-                        this.data.remoteLaserOrigin.y,
-                        this.data.remoteLaserOrigin.z
-                    )
-                );
-            }
-
-            if (prevData.laserTarget && !almostEquals(0.001, prevData.laserTarget, this.data.laserTarget)) {
-                this.targetBuffer.setPosition(
-                    targetBufferPosition.set(this.data.laserTarget.x, this.data.laserTarget.y, this.data.laserTarget.z)
-                );
-            }
-        };
-    })(),
-
-    tick: (() => {
-        const origin = new THREE.Vector3();
-        const target = new THREE.Vector3();
-        return function (t, dt) {
-            const isMine =
-                this.el.parentEl.components.networked.initialized && this.el.parentEl.components.networked.isMine();
-            let laserVisible = false;
-
-            if (isMine && this.data.laserVisible) {
-                origin.copy(this.data.laserOrigin);
-
-                if (!this.data.laserInHand) {
-                    // On 2d mode, shift downards
-                    origin.y = origin.y - 0.4;
-                }
-
-                target.copy(this.data.laserTarget);
-                laserVisible = true;
-            } else if (!isMine && this.data.laserVisible) {
-                this.originBuffer.update(dt);
-                this.targetBuffer.update(dt);
-                origin.copy(this.originBuffer.getPosition());
-                target.copy(this.targetBuffer.getPosition());
-                laserVisible = true;
-            }
-
-            if (laserVisible) {
-                this.laser.position.copy(origin);
-                this.laser.lookAt(target);
-                this.laser.scale.set(1, 1, origin.distanceTo(target));
-                this.laser.matrixNeedsUpdate = true;
-            }
-
-            if (this.laser.material.visible !== laserVisible) {
-                this.laser.material.visible = laserVisible;
-            }
-        };
-    })(),
-
-    remove() {
-        this.el.sceneEl.removeObject3D(`pen-laser-${this.laser.uuid}`);
-    }
-});
