+++ src/hub-components.js
@@ -6,8 +6,6 @@ import "./components/fader";
 import "./components/virtual-gamepad-controls";
 import "./components/ik-controller";
 import "./components/hand-controls2";
-import "./components/hoverable-visuals";
-import "./components/hover-visuals";
 import "./components/offset-relative-to";
 import "./components/player-info";
 import "./components/name-tag";
@@ -18,33 +16,17 @@ import "./components/visibility-on-content-types";
 import "./components/hide-when-pinned-and-forbidden";
 
 import "./mega-src/components/screenshot";
-import "./mega-src/components/students-can-move";
-import "./mega-src/components/locked";
-import "./mega-src/components/owner";
-import "./mega-src/components/auto-rotate";
-import "./mega-src/components/question";
-import "./mega-src/components/mega-image";
-import "./mega-src/components/open-link-button";
-import "./mega-src/components/description";
 import "./mega-src/systems/CSS2DRenderingSystem";
 // import "./mega-src/systems/CSS3DRenderingSystem";
 import "./mega-src/systems/tip-system";
 
 import "./components/networked-avatar";
-import "./components/media-video";
-// import "./components/media-pdf";
-import "./components/media-image";
 import "./components/pinch-to-move";
 import "./components/pitch-yaw-rotator";
 import "./components/position-at-border";
-import "./components/pinnable";
 
-import "./components/unmute-video-button";
-import "./components/visible-to-owner";
 
 import "./components/emit-scene-event-on-remove";
-import "./components/follow-in-fov";
-import "./components/hover-menu";
 import "./components/teleporter";
 import "./components/track-pose";
 import "./components/visibility-by-path";
@@ -55,15 +37,9 @@ import "./components/avatar-inspect-collider";
 import "./components/video-texture-target";
 import "./components/mirror";
 
-import "./components/event-repeater";
 import "./components/set-yxz-order";
 
 import "./components/cursor-controller";
-
-import "./components/tools/pen";
-import "./components/tools/pen-laser";
-import "./components/tools/networked-drawing";
-import "./components/tools/drawing-manager";
 
 import "./components/body-helper";
 import "./components/shape-helper";
